requirements.txt
# Phase 1: Database
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# Phase 2: Data Pipeline
yt-dlp>=2023.10.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
Pillow>=10.0.0
tqdm>=4.66.0

# Phase 3: Feature Engineering
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
open-clip-torch>=2.20.0

# Phase 4: Retrieval Model
faiss-cpu>=1.7.4
scikit-learn>=1.3.0